계산기 기능 구현 수정, 코드리뷰 반영
